Move order-by util into separate file
