+++ workers/loc.api/helpers/utils/index.js
@@ -1,0 +1,7 @@
+'use strict'
+
+const orderBy = require('./order-by')
+
+module.exports = {
+  orderBy
+}
